misc: hed-software-configurable-io: Add ADC input support
The existing driver was setup to just set the input selection pins
based on the mode set.  This addition adds reading the ADC that can
measure the input voltage and output the actual voltage in mV for VTD
or a 1 or 0 for STB and STG in the sysfs input_value file.

This also makes the input-gpio optional since some processors do not
support GPIO pins that are also ADCs on the same pin.