imx8mp-blizzard.dts: Change keyswitch debounce to 50
Change the keyswitch debounce interval to 50ms, the default is 5ms.

The reason behind this is that some customers have very noisy keyswitch
wiring. The GPIO reads based on TTL logic. If it is noisy, it can bounce
around close enough to one level or another, to average out to a high or
low signal. From testing, this happens around the 5ms or 10ms area. 50ms
is a compromise that seems to work well enough for everyone.