imx8mp-blizzard.dts: Change keyswitch to ACTIVE_LOW
Change the polarity of the keyswitch gpio from `GPIO_ACTIVE_HIGH` to
`GPIO_ACTIVE_LOW`. Due to the way it's wired, If it's
`GPIO_ACTIVE_HIGH`, and the keyswitch is plugged in (it needs to be
plugged in, in order to boot), then on bootup systemd think the "lid" is
closed, and takes the action in `HandleLidSwitch` in
`/etc/systemd/logind.conf`. We want the opposite, if it's plugged in,
take no action, if it's unplugged then take the action in
`HandleLidSwitch`.